+++ DataLayer/DataLayer.projitems
@@ -9,6 +9,9 @@
     <Import_RootNamespace>SchoolGrades.DataLayer</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)AnnotationData.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DataLayer.cs" />
+    <Compile Include="..\DataLayer\TestLessonsTopicAndQuestionsData.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)AnswersGradesAndStudentData.cs" />
   </ItemGroup>
 </Project>
