+++ DataLayer/TestLessonsAndQuestionsData.cs
@@ -1,0 +1,445 @@
+﻿using System;
+using System.Data.Common;
+using System.Data;
+using System.Data.SQLite;
+using System.IO;
+using System.Collections.Generic;
+using gamon;
+using System.Windows.Forms;
+using SchoolGrades.DbClasses;
+
+namespace SchoolGrades.DataLayer
+{
+    class TestLessonTopicsAndQuestionsData
+    {
+        DataLayer dl = new DataLayer();
+
+        internal Test GetTestFromRow(DbDataReader Row)
+        {
+            Test t = new Test();
+            t.IdTest = SafeDb.SafeInt(Row["idTest"]);
+            t.Name = SafeDb.SafeString(Row["name"]);
+            t.Desc = SafeDb.SafeString(Row["desc"]);
+            t.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
+            t.IdTestType = SafeDb.SafeString(Row["IdTestType"]);
+            t.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
+            t.IdTopic = SafeDb.SafeInt(Row["IdTopic"]);
+
+            return t;
+        }
+        internal Test GetTest(int? IdTest)
+        {
+            Test t = new Test();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * " +
+                    "FROM Tests " +
+                    "WHERE IdTest=" + IdTest +
+                    ";";
+                dRead = cmd.ExecuteReader();
+
+                while (dRead.Read())
+                {
+                    t = GetTestFromRow(dRead);
+                }
+            }
+            return t;
+        }
+        internal List<Test> GetTests()
+        {
+            List<Test> list = new List<Test>();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * " +
+                    "FROM Tests " +
+                    //"WHERE idSchoolYear=" + IdSchoolYear +
+                    //" OR IdSchoolYear IS null OR IdSchoolYear=''" +
+                    ";";
+                dRead = cmd.ExecuteReader();
+
+                while (dRead.Read())
+                {
+                    Test t = GetTestFromRow(dRead);
+                    list.Add(t);
+                }
+            }
+            return list;
+        }
+
+        internal void SaveTest(Test TestToSave)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                if (TestToSave.IdTest == 0 || TestToSave.IdTest == null)
+                {   // create new record
+                    int nextId = NextKey("Tests", "idTest");
+
+                    cmd.CommandText = "INSERT INTO Tests " +
+                    "(idTest,name,desc,IdSchoolSubject,IdTestType,IdTopic" +
+                    ")" +
+                    "Values " +
+                    "(" + nextId + ",'" + SqlVal.SqlString(TestToSave.Name) + "','" +
+                    SqlVal.SqlString(TestToSave.Desc) + "','" + SqlVal.SqlString(TestToSave.IdSchoolSubject) + "'," +
+                     SqlVal.SqlInt(TestToSave.IdTestType) + "," + SqlVal.SqlInt(TestToSave.IdTopic) +
+                    ");";
+                }
+                else
+                {   // update old record
+                    cmd.CommandText = "UPDATE Tests" +
+                    " SET name='" + SqlVal.SqlString(TestToSave.Name) + "'," +
+                    "desc='" + SqlVal.SqlString(TestToSave.Desc) + "'" +
+                    ",IdSchoolSubject=" + SqlVal.SqlString(TestToSave.IdSchoolSubject) +
+                    ",IdTestType=" + SqlVal.SqlInt(TestToSave.IdTestType) +
+                    ",IdTopic=" + SqlVal.SqlInt(TestToSave.IdTopic) +
+                    ")" +
+                    " WHERE idTest=" + TestToSave.IdTest +
+                    ";";
+                }
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+        internal List<Question> GetAllQuestionsOfATest(int? IdTest)
+        {
+            List<Question> lq = new List<Question>();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT *" +
+                    " FROM Questions" +
+                    " JOIN Tests_Questions ON Tests_Questions.IdQuestion=Questions.IdQuestion" +
+                    " WHERE Tests_Questions.idTest=" + IdTest +
+                    ";";
+                DbDataReader dRead;
+                dRead = cmd.ExecuteReader();
+
+                while (dRead.Read())
+                {
+                    Question q = GetQuestionFromRow(dRead);
+                    lq.Add(q);
+                }
+            }
+            return lq;
+        }
+        internal void AddQuestionToTest(Test Test, Question Question)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                // get the code of the previous photo
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Tests_Questions" +
+                    " (IdTest, IdQuestion)" +
+                    " Values" +
+                    " (" + Test.IdTest + "," + Question.IdQuestion + ")" +
+                    "; ";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+        internal List<QuestionType> GetListQuestionTypes(bool IncludeANullObject)
+        {
+            List<QuestionType> l = new List<QuestionType>();
+            if (IncludeANullObject)
+            {
+                QuestionType qt = new QuestionType();
+                qt.IdQuestionType = "";
+                l.Add(qt);
+            }
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * FROM QuestionTypes;";
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    QuestionType o = new QuestionType();
+                    o.Name = (string)dRead["Name"];
+                    o.IdQuestionType = (string)dRead["IdQuestionType"];
+                    o.Desc = (string)dRead["Desc"];
+                    o.IdQuestionType = (string)dRead["IdQuestionType"];
+                    l.Add(o);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+                return l;
+            }
+        }
+
+        internal int NewLesson(Lesson Lesson)
+        {
+            int key;
+            using (DbConnection conn = dl.Connect())
+            {
+                key = NextKey("Lessons", "idLesson");
+                Lesson.IdLesson = key;
+                // add new record to Lessons table
+                DbCommand cmd = conn.CreateCommand();
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Lessons" +
+                " (idLesson, date, idClass, idSchoolSubject, idSchoolYear, note) " +
+                "Values (" +
+                "" + Lesson.IdLesson + "" +
+                "," + SqlVal.SqlDate(Lesson.Date) + "" +
+                "," + Lesson.IdClass + "" +
+                ",'" + Lesson.IdSchoolSubject + "'" +
+                ",'" + Lesson.IdSchoolYear + "'" +
+                ",'" + SqlVal.SqlString(Lesson.Note) + "'" +
+                ");";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+            return key;
+        }
+        internal void SaveLesson(Lesson Lesson)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "UPDATE Lessons" +
+                " SET" +
+                " date=" + SqlVal.SqlDate(Lesson.Date) + "," +
+                " idClass=" + Lesson.IdClass + "," +
+                " idSchoolSubject='" + Lesson.IdSchoolSubject + "'," +
+                " idSchoolYear='" + Lesson.IdSchoolYear + "'," +
+                " note='" + SqlVal.SqlString(Lesson.Note) + "'" +
+                " WHERE idLesson=" + Lesson.IdLesson +
+                ";";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+        }
+        internal object GetTopicsOfOneLessonOfClass(Class Class, Lesson Lesson)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT Topics.* FROM Lessons" +
+                    " LEFT JOIN Lessons_Topics ON Lessons.IdLesson = Lessons_Topics.idLesson" +
+                    " LEFT JOIN Topics ON Topics.idTopic = Lessons_Topics.idTopic" +
+                    " WHERE idSchoolSubject='" + Lesson.IdSchoolSubject + "'" +
+                    " AND Lessons.idSchoolYear='" + Lesson.IdSchoolYear + "'" +
+                    " AND Lessons.idClass='" + Class.IdClass + "'" +
+                    " AND Lessons.idLesson='" + Lesson.IdLesson + "'" +
+                    //" GROUP BY Lessons.idLesson" +
+                    " ORDER BY Lessons.date DESC" +
+                    ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetOnLessonOfClass");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dAdapt.Dispose();
+                dSet.Dispose();
+            }
+            return t;
+        }
+        internal DataTable GetLessonsOfClass(Class Class, Lesson Lesson)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT * FROM Lessons" +
+                    " WHERE idSchoolSubject='" + Lesson.IdSchoolSubject + "'" +
+                    " AND Lessons.idSchoolYear='" + Lesson.IdSchoolYear + "'" +
+                    " AND Lessons.idClass='" + Class.IdClass + "'" +
+                    //" GROUP BY Lessons.idLesson" +
+                    " ORDER BY Lessons.date DESC" +
+                    ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetLessonsOfClass");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dAdapt.Dispose();
+                dSet.Dispose();
+            }
+            return t;
+        }
+        internal Lesson GetLastLesson(Lesson CurrentLesson)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                List<Couple> couples = new List<Couple>();
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                Lesson l;
+                string query;
+                query = "SELECT * FROM Lessons" +
+                        " WHERE idClass=" + CurrentLesson.IdClass.ToString() +
+                        " AND idSchoolSubject='" + CurrentLesson.IdSchoolSubject + "'" +
+                        " AND idSchoolYear='" + CurrentLesson.IdSchoolYear + "'" +
+                        " ORDER BY Date DESC LIMIT 1;";
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                l = new Lesson();
+                while (dRead.Read())
+                {
+                    l.IdLesson = (int)dRead["idLesson"];
+                    l.Date = (DateTime)dRead["Date"];
+                    l.IdClass = (int)dRead["idClass"];
+                    l.IdSchoolSubject = (string)dRead["idSchoolSubject"];
+                    l.Note = (string)dRead["note"];
+
+                    break;
+                }
+                cmd.Dispose();
+                dRead.Dispose();
+                return l;
+            }
+        }
+        internal Lesson GetLessonInDate(Class Class, string IdSubject,
+            DateTime Date)
+        {
+            Lesson less = new Lesson();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                string query;
+                query = "SELECT * FROM Lessons" +
+                        " WHERE idClass=" + Class.IdClass.ToString() +
+                        " AND idSchoolSubject='" + IdSubject + "'" +
+                        " AND date BETWEEN " + SqlVal.SqlDate(Date.ToString("yyyy-MM-dd")) +
+                        " AND " + SqlVal.SqlDate(Date.AddDays(1).ToString("yyyy-MM-dd")) +
+                        " LIMIT 1;";
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    less.IdLesson = (int)dRead["idLesson"];
+                    less.Date = (DateTime)dRead["Date"];
+                    less.IdClass = (int)dRead["idClass"];
+                    less.IdSchoolSubject = (string)dRead["idSchoolSubject"];
+                    less.Note = (string)dRead["note"];
+
+                    break; // there should be only one record in the query result 
+                }
+                cmd.Dispose();
+                dRead.Dispose();
+            }
+            return less;
+        }
+        internal void EraseLesson(int? IdLesson, bool AlsoEraseImageFiles)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                // erase existing links topic-lesson
+                DbCommand cmd = conn.CreateCommand();
+                string query = "DELETE FROM Lessons_Topics" +
+                        " WHERE idLesson=" + IdLesson.ToString() +
+                        ";";
+                cmd.CommandText = query;
+                cmd.ExecuteNonQuery();
+
+                // erase images' files if permitted
+                if (AlsoEraseImageFiles)
+                {
+                    // !! TODO !! find the images that aren't linked to another lesson and delete
+                    // the files if EraseImageFiles is set 
+                    throw new NotImplementedException();
+                }
+
+                // erase existing links images-lesson
+                cmd = conn.CreateCommand();
+                query = "DELETE FROM Lessons_Images" +
+                        " WHERE idLesson=" + IdLesson.ToString() +
+                        ";";
+                cmd.CommandText = query;
+                cmd.ExecuteNonQuery();
+
+                // erase the lesson's row from lessons
+                cmd = conn.CreateCommand();
+                query = "DELETE FROM Lessons" +
+                        " WHERE idLesson=" + IdLesson.ToString() +
+                        ";";
+                cmd.CommandText = query;
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+        internal void SaveTopicsOfLesson(int? IdLesson, List<Topic> topicsOfTheLesson)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                // erase existing links topic-lesson
+                DbCommand cmd = conn.CreateCommand();
+                string query = "DELETE FROM Lessons_Topics" +
+                        " WHERE idLesson=" + IdLesson.ToString() +
+                        ";";
+                cmd.CommandText = query;
+                cmd.ExecuteNonQuery();
+                int insertionOrder = 1;
+                foreach (Topic t in topicsOfTheLesson)
+                {
+                    // insert links topic-lesson, one at a time 
+                    cmd.CommandText = "INSERT INTO Lessons_Topics" +
+                    " (idLesson, idTopic, insertionOrder)" +
+                    " Values (" +
+                    "" + IdLesson + "" +
+                    "," + t.Id +
+                    "," + insertionOrder++ +
+                    ");";
+                    try
+                    {
+                        cmd.ExecuteNonQuery();
+                    }
+
+                    catch (Exception e)
+                    {
+                        {
+                            // if it isn't UNIQUE, then it is error (to cure!!!!)
+                            // but actually it doesn't matter too much, because
+                            // the lesson is already linked to the topic anyway.. 
+
+
+                        }
+                    }
+                }
+                cmd.Dispose();
+            }
+        }
+        internal List<Topic> GetTopicsOfLesson(int? IdLesson, List<Topic> topicsOfTheLesson)
+        {
+            if (IdLesson == null)
+            {
+                return null;
+            }
+            // order by ensures that the order of the result is the order of insertion 
+            // in the database (that was the same of the tree traversal) 
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                DbDataReader dRead;
+                string query;
+                query = "SELECT * FROM Topics" +
+                        " JOIN Lessons_Topics ON Topics.idTopic=Lessons_Topics.idTopic" +
+                        " WHERE Lessons_Topics.idLesson=" + IdLesson +
+                        " ORDER BY insertionOrder" +
+                        ";";
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    Topic t = GetTopicFromRow(dRead);
+                    topicsOfTheLesson.Add(t);
+                }
+                cmd.Dispose();
+                dRead.Dispose();
+            }
+            return topicsOfTheLesson;
+        }
+    }
+}
